Refactor models and admin for better organization
