+++ rounds/admin.py
@@ -1,3 +1,6 @@
 from django.contrib import admin
+from .models import Round, HaiPai, WinningHand
 
-# Register your models here.
+admin.site.register(Round)
+admin.site.register(HaiPai)
+admin.site.register(WinningHand)
